Add support for Rust code built through Cargo
This recognizes the .rs extension, and extends the pattern matching to
both recognize Rust symbols (with their mangled names, no demangling in
here yet) and to recognize paths stemming from Rust's build system. (One
path component is stripped as it represents the overall repository a
packet comes from, currently usually "github", and is not really helpful
to the user).